// src/generated/models/DetectRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DetectRequest = {
    url?: (string | null);
    source_id?: (string | null);
    page_range?: (string | null);
    include_symbols?: boolean;
    filter_pattern?: (string | null);
};