// src/generated/models/EngineeringExtractRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EngineeringExtractRequest = {
    url?: (string | null);
    source_id?: (string | null);
    page_range?: (string | null);
    accuracy_mode?: string;
};