// src/generated/models/ExtractRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExtractRequest = {
    url?: (string | null);
    source_id?: (string | null);
    page_range?: (string | null);
    json_schema: Record<string, any>;
};