// src/generated/models/ExtractResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExtractResponse = {
    id: string;
    status: string;
    progress?: number;
    url?: (string | null);
    source_display_name?: (string | null);
    page_range?: (string | null);
    records?: Array<Record<string, any>>;
    record_count?: number;
    page_count?: number;
    error?: (string | null);
    created_at: string;
    updated_at: string;
};